// apps/web/src/api/models/ruleResponse.ts
/**
 * Generated by orval v7.17.2 🍺
 * Do not edit manually.
 * PayPay Money Diary API
 * PayPay 家計簿アプリケーション API
 * OpenAPI spec version: 1.0.0
 */

export interface RuleResponse {
  id: string;
  /** @nullable */
  userId: string | null;
  keyword: string;
  categoryId: string;
  /** @nullable */
  categoryName: string | null;
  priority: number;
  isSystem: boolean;
  createdAt: string;
  updatedAt: string;
}